+++ app/src/main/java/com/example/voyagerx/ui/fragments/landing/data/LaunchOverviewData.kt
@@ -1,3 +1,3 @@
-package com.example.voyagerx.ui.fragments.landing
+package com.example.voyagerx.ui.fragments.landing.data
 
 data class LaunchOverviewData(val missionName: String, val siteName: String, val year: String)
