+++ app/src/main/java/com/example/voyagerx/ui/fragments/landing/list/LaunchOverviewClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.voyagerx.ui.fragments.landing.list
 
-import com.example.voyagerx.ui.fragments.landing.LaunchOverviewData
+import com.example.voyagerx.ui.fragments.landing.data.LaunchOverviewData
 
 class LaunchOverviewClickListener(
     private val listener: (launchOverviewData: LaunchOverviewData) -> Unit
