+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,9 @@
 <resources>
     <string name="app_name">VoyagerX</string>
 
-    <string name="example_mission_name">FalconSat</string>
-    <string name="example_launch_overview_year">2006</string>
-    <string name="example_launch_overview_site_name">Kwajalein Atoll Omelek Island</string>
+    <string name="example_mission_name">SXM-7</string>
+    <string name="example_launch_overview_year">2020</string>
+    <string name="example_launch_overview_site_name">Cape Canaveral Air Force Station Space Launch Complex 40</string>
     <string name="field_separator">•</string>
     <string name="filter_launches_content_description">Filter launches</string>
     <string name="launch_listing_header">There are %1$d total launches 🚀</string>
